+++ app/models/concerns/recording_studio_user/profiled_user.rb
@@ -8,5 +8,9 @@ module RecordingStudioUser
       validates :first_name, :last_name, :time_zone, presence: true
       validates :time_zone, inclusion: { in: ->(_record) { ActiveSupport::TimeZone.all.map(&:name) } }
     end
+
+    def display_name
+      RecordingStudioUser.display_name_for(self)
+    end
   end
 end